# docs/swagger.yaml
info:
  contact: {}
paths:
  /api/v1/clouds/node/:node_name/containers/create:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 创建容器
      tags:
      - 容器
  /api/v1/clouds/node/:node_name/containers/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 获取所有容器列表
      tags:
      - 容器
  /api/v1/clouds/node/:node_name/containers/log:
    post:
      parameters:
      - description: 容器名称
        in: path
        name: containerName
        required: true
        type: string
      - description: 容器ID
        in: path
        name: containerId
        required: true
        type: string
      - description: 是否监听日志
        in: path
        name: isWatch
        type: boolean
      - description: 获取日志范围, all,1m,10m...
        in: path
        name: mode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 容器日志
      tags:
      - 容器
  /api/v1/clouds/node/:node_name/containers/option:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 容器操作选项
      tags:
      - 容器
  /api/v1/clouds/node/:node_name/containers/search:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 获取指定容器
      tags:
      - 容器
  /api/v1/clouds/node/:node_name/containers/state/:container_id:
    get:
      parameters:
      - description: 容器ID
        in: path
        name: containerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 获取容器资源
      tags:
      - 容器
  /api/v1/clouds/node/:node_name/containers/state/:id:
    get:
      parameters:
      - description: 容器ID
        in: path
        name: containerId
        required: true
        type: string
      - description: user
        in: path
        name: user
        required: true
        type: string
      - description: 执行命令
        in: path
        name: command
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 容器终端
      tags:
      - 容器
  /api/v1/clouds/node/:node_name/images/delete:
    post:
      parameters:
      - description: 镜像名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 删除镜像
      tags:
      - 镜像
  /api/v1/clouds/node/:node_name/images/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 获取所有镜像
      tags:
      - 镜像
  /api/v1/clouds/node/:node_name/images/pull:
    post:
      parameters:
      - description: 镜像名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 拉取镜像
      tags:
      - 镜像
  /api/v1/clouds/node/:node_name/images/search:
    post:
      parameters:
      - description: 镜像名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 搜索镜像
      tags:
      - 镜像
  /api/v1/clouds/node/:node_name/networks/create:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 创建网络
      tags:
      - 网络
  /api/v1/clouds/node/:node_name/networks/delete:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 删除网络
      tags:
      - 网络
  /api/v1/clouds/node/:node_name/networks/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 获取所有网络列表
      tags:
      - 网络
  /api/v1/clouds/node/:node_name/networks/search:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 搜索网络
      tags:
      - 网络
  /api/v1/clouds/node/:node_name/volumes/create:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 创建卷
      tags:
      - 卷
  /api/v1/clouds/node/:node_name/volumes/delete:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 删除卷
      tags:
      - 卷
  /api/v1/clouds/node/:node_name/volumes/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 获取所有卷
      tags:
      - 卷
  /api/v1/clouds/node/:node_name/volumes/search:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"请求成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"请求失败"}'
          schema:
            type: string
      summary: 搜索卷
      tags:
      - 卷
  /api/v1/clouds/node/create:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"创建成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"创建失败"}'
          schema:
            type: string
      summary: 创建 Docker 节点
      tags:
      - 节点
  /api/v1/clouds/node/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"删除成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"删除失败"}'
          schema:
            type: string
      summary: 删除 Docker 节点
      tags:
      - 节点
  /api/v1/clouds/node/update:
    put:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"更新成功"}'
          schema:
            type: string
        "400":
          description: '{"code":400,"data":{},"msg":"更新失败"}'
          schema:
            type: string
      summary: 更新 Docker 节点
      tags:
      - 节点
swagger: "2.0"